+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Correctly handle optional `x-okapi-token` request header. Fixes UIPCIR-61.
 * Remove not needed Paneset to correctly calculate pane widths when closing plugin. Fixes UIPCIR-78.
+* Jest/RTL: Cover `useOptions` hook with unit tests. Refs UIPCIR-68.
 * Jest/RTL: Cover `CreateRecordsWrapper` component with unit tests. Refs UIPCIR-64.
 
 ## [4.0.0](https://github.com/folio-org/ui-plugin-create-inventory-records/tree/v4.0.0) (2023-10-12)